+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * [#93](https://github.com/antoinezanardi/werewolves-assistant-web/issues/93) - Enhance Game Lobby display for small screens.
 * [#98](https://github.com/antoinezanardi/werewolves-assistant-web/issues/98) - Extend player limit.
+* [#109](https://github.com/antoinezanardi/werewolves-assistant-web/issues/109) - Add limits to waking up interval options.
 
 ### 🐛 Bug fixes